+++ lib/fog/softlayer/ext/hash.rb
@@ -1,0 +1,6 @@
+class Hash
+  def deep_merge(second)
+    merger = proc { |key, v1, v2| Hash === v1 && Hash === v2 ? v1.merge(v2, &merger) : v2 }
+    self.merge(second, &merger)
+  end
+end
