+++ lib/fog/softlayer/models/compute/key_pair.rb
@@ -5,8 +5,6 @@
 # LICENSE: MIT (http://opensource.org/licenses/MIT)
 #
 
-require 'fog/core/model'
-
 module Fog
   module Compute
     class Softlayer
