+++ lib/fog/softlayer/models/compute/network_component.rb
@@ -5,8 +5,6 @@
 # LICENSE: MIT (http://opensource.org/licenses/MIT)
 #
 
-require 'fog/core/model'
-
 module Fog
   module Compute
     class Softlayer
