+++ lib/fog/softlayer/models/dns/domain.rb
@@ -4,7 +4,7 @@
 #
 # LICENSE: MIT (http://opensource.org/licenses/MIT)
 #
-require 'fog/core/model'
+
 require 'fog/softlayer/models/dns/records'
 
 module Fog
