+++ lib/fog/softlayer/models/dns/domains.rb
@@ -4,7 +4,7 @@
 #
 # LICENSE: MIT (http://opensource.org/licenses/MIT)
 #
-require 'fog/core/collection'
+
 require 'fog/softlayer/models/compute/server'
 
 module Fog
