+++ lib/fog/softlayer/models/dns/records.rb
@@ -4,7 +4,7 @@
 #
 # LICENSE: MIT (http://opensource.org/licenses/MIT)
 #
-require 'fog/core/collection'
+
 require 'fog/softlayer/models/dns/record'
 
 module Fog
