+++ lib/fog/softlayer/models/network/datacenter.rb
@@ -5,8 +5,6 @@
 # LICENSE: MIT (http://opensource.org/licenses/MIT)
 #
 
-require 'fog/core/model'
-
 module Fog
   module Network
     class Softlayer
