+++ lib/fog/softlayer/models/network/tags.rb
@@ -4,7 +4,7 @@
 #
 # LICENSE: MIT (http://opensource.org/licenses/MIT)
 #
-require 'fog/core/collection'
+
 require 'fog/softlayer/models/network/tag'
 
 module Fog
