+++ lib/fog/softlayer/storage.rb
@@ -5,8 +5,6 @@
 # LICENSE: MIT (http://opensource.org/licenses/MIT)
 #
 
-require 'fog/softlayer/core'
-
 module Fog
   module Storage
     class Softlayer < Fog::Service
